+++ apps/qna/services/answer/command.py
@@ -8,7 +8,7 @@ from google.genai import types
 from rest_framework import status
 
 from apps.qna.constants import ErrorMessages
-from apps.qna.exceptions.base import QnaBaseException
+from apps.qna.exceptions import QnaBaseException
 from apps.qna.models import (
     Answer,
     AnswerComment,
